Compile fixes systems with uapi/linux/bpf.h include file below kernel v4.13
This bpf-tool project doesn't maintain an UAPI bpf.h include
file, but depend on it being installed on the system distro.
(Fedora provide this uapi/linux/bpf.h in package: kernel-devel)

When needing/depending some define or struct that is only
available in a newer version, this leads to strange compile errors.

This bpf-tool depend on features that were introduced in kernel
v4.13, which hardly any distros provide (given in was released 12
days ago, as of this writing).

Generally the BPF UAPI use enum's in the UAPI. Thus, the usual
#ifdef tricks doesn't work, as they are not expanded during the
C-preprocessor compiler step.

Fortunately we got lucky, as the struct needed by this tool, also
moved the define BPF_TAG_SIZE into UAPI.  Thus, use this to
detect if struct bpf_prog_info and struct bpf_map_info need to be
defined locally.

Dependency is on kernel:
 commit 1e2709769086 ("bpf: Add BPF_OBJ_GET_INFO_BY_FD")
 https://git.kernel.org/torvalds/c/1e2709769086

Signed-off-by: Jesper Dangaard Brouer <brouer@redhat.com>